pyproject.toml
[tool.poetry]
name = "creatio-api-py"
version = "1.0.0"
description = "Python client for Creatio OData API"
license = "GPL-3.0-only"
authors = ["Alejandro González Momblán <alejandro.gonzalez.momblan@evoluciona.es>"]
readme = "README.md"
repository = "https://github.com/yisuschrist/creatio-api-py"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
rich = "^13.7.0"
requests-cache = "^1.1.1"
requests-pprint = "^1.0.0"

[tool.poetry.group.dev.dependencies]
bandit = "^1.7.7"
black = "^24.2.0"
mypy = "^1.8.0"
pre-commit = "^3.6.1"
pylint = "^3.0.3"
pytest = "^8.0.0"
pyupgrade = "^3.15.0"
coverage = "^7.4.1"

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["creatio_odata_api", "tests"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
implicit_reexport = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
